requirements.txt
feedparser
chromadb
sentence-transformers
transformers
torch
google-generativeai
spacy
flask
flask-cors
python-dotenv
